Change options to avoid conflicts with new Django 1.6 options
Django 1.6 adds -t/--top-level-directory and -p/--pattern. We were using -t and -p which resulted in optparse.OptionConflictError when trying to use test_extras with Django 1.6.

Renamed -t to -g (wanted to keep a short option for --tags because it is commonly used).

Removed -p short option, use long --pdb option instead (I expect that this option is less commonly used than -t/--tags).

Bump version to 1.1.0 because these changes are not backwards compatible - scripts using -t and -p will need to be changed.

Add tox environments for Django 1.5 and 1.6.